create blog layout FE
